chore(build): enable local macOS test gate
qlean is only used by QEMU/KVM integration tests that CI never compiles
(`--lib --bins`), but it unconditionally pulls kvm-ioctls which does not
compile on macOS arm64. Gate it to linux/x86_64 so `cargo test --lib
--bins` (the repo's required local CI gate) can run on macOS.

Also fix test-only portability: libc::makedev and S_IFIFO/S_IFCHR use
different integer widths on darwin, so cast them explicitly.

+++ src/fuse/mod.rs
@@ -3761,7 +3761,7 @@ mod fuse_init_tests {
             Request::default(),
             1,
             OsStr::new("pipe"),
-            libc::S_IFIFO | 0o644,
+            (libc::S_IFIFO as u32) | 0o644,
             0,
         )
         .await
@@ -3784,7 +3784,7 @@ mod fuse_init_tests {
             Request::default(),
             1,
             OsStr::new("tty"),
-            libc::S_IFCHR | 0o600,
+            (libc::S_IFCHR as u32) | 0o600,
             0x0103,
         )
         .await
